Switch to CarrierWave for uploads and process uploads in background.
The switch to CarrierWave is so that we can pre-process the different
versions of the file (rather than them being built on demand in refile).

We've also changed things so that we pre-process all of the image
uploads asynchronously with delayed_job. This should let file uploads
work even if it would normally take a long time to process.

These changes were made to try to workaround large file uploads not
working out on Heroku.